Create test file for _store_backends module
