+++ joblib/test/test_memory.py
@@ -13,14 +13,6 @@ import pickle
 import sys
 import time
 import datetime
-try:
-    # Python 2.7: use the C pickle to speed up
-    # test_concurrency_safe_write which pickles big python objects
-    import cPickle as cpickle
-except ImportError:
-    import pickle as cpickle
-import functools
-
 
 from joblib.memory import Memory
 from joblib.memory import MemorizedFunc, NotMemorizedFunc
@@ -33,10 +25,8 @@ from joblib.parallel import Parallel, delayed
 from joblib._store_backends import StoreBackendBase
 from joblib.test.common import with_numpy, np
 from joblib.test.common import with_multiprocessing
-from joblib.testing import parametrize, raises, warns, timeout
+from joblib.testing import parametrize, raises, warns
 from joblib._compat import PY3_OR_LATER
-from joblib.backports import concurrency_safe_rename
-from joblib._store_backends import concurrency_safe_write
 
 
 ###############################################################################
@@ -821,48 +811,6 @@ def test_cached_function_race_condition_when_persisting_output_2(tmpdir,
     exception_msg = 'Exception while loading results'
     assert exception_msg not in stdout
     assert exception_msg not in stderr
-
-
-def write_func(output, filename):
-    with open(filename, 'wb') as f:
-        cpickle.dump(output, f)
-
-
-def concurrency_safe_write_rename(to_write, filename, write_func):
-    temporary_filename = concurrency_safe_write(to_write,
-                                                filename, write_func)
-    concurrency_safe_rename(temporary_filename, filename)
-
-
-def load_func(expected, filename):
-    for i in range(10):
-        try:
-            with open(filename, 'rb') as f:
-                reloaded = cpickle.load(f)
-            break
-        except (OSError, IOError):
-            # On Windows you can have WindowsError ([Error 5] Access
-            # is denied or [Error 13] Permission denied) when reading the file,
-            # probably because a writer process has a lock on the file
-            time.sleep(0.1)
-    else:
-        raise
-    assert expected == reloaded
-
-
-@timeout(0)  # No timeout as this test can be long
-@with_multiprocessing
-@parametrize('backend', ['multiprocessing', 'loky', 'threading'])
-def test_concurrency_safe_write(tmpdir, backend):
-    # Add one item to cache
-    filename = tmpdir.join('test.pkl').strpath
-
-    obj = {str(i): i for i in range(int(1e5))}
-    funcs = [functools.partial(concurrency_safe_write_rename,
-                               write_func=write_func)
-             if i % 3 != 2 else load_func for i in range(12)]
-    Parallel(n_jobs=2, backend=backend)(
-        delayed(func)(obj, filename) for func in funcs)
 
 
 def test_memory_recomputes_after_an_error_why_loading_results(tmpdir,
